disable mobile view || change PayrollsModel paymentamount and payableamount to long datatype to fix arithmetic error || fix index form layout

+++ CAJWebApp/Areas/PAYROLL/Models/PayrollsModel.cs
@@ -260,11 +260,11 @@ namespace CAJWebApp.Areas.PAYROLL.Models
 
 
         [DisplayFormat(DataFormatString = "{0:N0}")]
-        public int PayableAmount { get; set; }
+        public long PayableAmount { get; set; }
 
 
         [DisplayFormat(DataFormatString = "{0:N0}")]
-        public int PaymentAmount { get; set; }
+        public long PaymentAmount { get; set; }
